requirements.txt
# RICK Trading System - Core Dependencies
# Broker Connectors
oandapyV20>=0.7.2
coinbase-advanced-py>=1.4.0
ib_insync>=0.9.86
websocket-client>=1.6.0

# Core Python Libraries
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Data Analysis & ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Scheduling / task runner
apscheduler>=3.10.4

# Logging (structured)
structlog>=24.1.0

# Visualization & Dashboard
streamlit>=1.28.0
graphviz>=0.20.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
